User case 1 & 2 ok

+++ app/src/resolver/rating_schema.py
@@ -5,7 +5,7 @@ import strawberry
 @strawberry.input
 class PlayerRatingInputType:
     player_id: int = strawberry.field(name="player_id")
-    player_average_rating: int = strawberry.field(name="player_average_rating")
+    player_average_rating: float = strawberry.field(name="player_average_rating")
 
 
 @strawberry.input
